srcs/backend/test-service/requirements.txt
django
djangorestframework #django rest framework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
psycopg2-binary  # postgres

django-cors-headers # cors
django-prometheus #prometheus monitoring
django-watchman #health check status
